+++ NEWS.md
@@ -1,6 +1,7 @@
 # branch no-phangorn (9002)
 
-- `PathDist()` uses internal implementation of path distance.
+- Internal implementation of path and SPR distances, removing dependency
+  on phangorn (and thus R 4.4).
 
 
 # TreeDist 2.7.1 (2024-06-13)